Add files via upload
processes.txt file contains a list of processes with pertinent data for the process.
Each line will have a process name that will be a string. Following that will be the arrival time of the process. The arrival time will be in reference to the previous process. Following this will be the total execution time. Next will be the elapsed time between I/O interrupts, next will be the time spent waiting and processing the I/O